Refactor: move processing logic to module
